requirements.txt
numpy==1.12.0
Pillow==4.1.0
Flask==0.12
tensorflow
matplotlib==1.5.3
scipy==0.18.0
pandas==0.18.1
SharedArray==1.0
click==6.6
scikit-image==0.12.3
scikit-learn==0.18.1
six==1.10.0
ghalton==0.6
psutil==5.0.0
gunicorn==19.7.0